resume commit
@foxelcode did it all
